requirements.txt
langchain
PyPDF2
sentence-transformers
faiss-gpu
transformers
torch
optimum
auto-gptq
